Fix a bug with sorting & uniquifying when using a locale
The locale comparer can return equal for strings which are different. For
example, when sorting with `en-US` the following strings are equal:

    A
    a
    ª

This means that the sorter would simply keep these in their original
order. That means that the sorted output would depend on the original order,
meaning two files with different orderings but the same lines would sort
differently.

This also broke the `--unique` flag because the `lines.dedup_by` call relies
on duplicate lines being adjacent. If they're not it won't properly dedup the
`Vec`.